Make Namespaces associated with an OpEnvironment
This creates a common interface, Environmental, for things with an
associated OpEnvironment. Both Op and Namespace now implement that
interface, giving each an associated execution environment, which
is accessible from subclasses via the ops() accessor method.

Namespaces are no longer singletons, but rather vanilla plugins,
so that they can be created on demand and tied to a particular
OpEnvironment. Therefore, the NamespaceService is a PTService,
not a SingletonService anymore.
